// src/app/Interfaces/istudentUpdate.ts
export default interface istudentUpdate {
    id?: string;
    userName?: string;
    email?: string;
    phoneNumber?: string;
    classYear?: string;
    age?: number;
    nationality?: string;
    iqamaNumber?: string;
    passportNumber?: string;
    classId?: string | null;
    curriculumId?: string;
}